Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,8 +1,8 @@
 #include<stdio.h>
 int main(){
-    int a,b,prod;
-    scanf("%d %d",&a,&b);
+    float a,b,prod;
+    scanf("%f %f",&a,&b);
     prod=a*b;
-    printf("Product: %d\n",prod);
+    printf("Product: %f\n",prod);
     return 0;
 }
